single origin script

+++ origin.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
-sbtVersion="0.11.2"
+if [ ! -e "home/interactive-shell.sh" ]
+then
+  sbtVersion="0.11.2"
+  git clone git@github.com:max-l/env-setup.git home
+  curl -s -o `pwd`/home/bin/sbt-launch.jar "http://repo.typesafe.com/typesafe/ivy-releases/org.scala-tools.sbt/sbt-launch/$sbtVersion/sbt-launch.jar"
+  echo "cmd.exe /c sh.exe home/interactive-shell.sh" > ./git-bash.cmd
+fi
 
-git clone git@github.com:max-l/env-setup.git home
-
-curl -s -o `pwd`/home/bin/sbt-launch.jar "http://repo.typesafe.com/typesafe/ivy-releases/org.scala-tools.sbt/sbt-launch/$sbtVersion/sbt-launch.jar"
-
-echo "cmd.exe /c sh.exe home/interactive-shell.sh" > ./git-bash.cmd
+home/interactive-shell.sh
